main/src/Applied_Analysis_Source/資産分類/requirement.txt
pandas>=1.3.0
openpyxl>=3.0.7
scikit-learn>=0.24.2
transformers>=4.5.1
torch>=1.9.0
numpy>=1.19.5
chardet>=4.0.0
pyyaml>=5.4.1
openai>=1.0.0
anthropic>=0.5.0
tenacity>=8.0.0
pymongo>=4.0.0
psycopg2-binary>=2.9.0
azure-storage-blob>=12.0.0
boto3>=1.26.0
google-cloud-storage>=2.0.0
